refactor(dns): promote dns module to directory
